Feat: Support f64

+++ src/data.rs
@@ -1,15 +1,15 @@
-#[derive(Debug, Eq, PartialEq)]
+#[derive(Debug, PartialEq)]
 pub enum ValueType {
     Null,
     Bool(bool),
-    Number(isize),
+    Number(f64),
     String(String),
     Array,
     // TODO: Vec<T>?
     Object, // TODO: inner?
 }
 
-#[derive(Debug, Eq, PartialEq)]
+#[derive(Debug, PartialEq)]
 pub struct Value {
     pub(crate) v_type: ValueType,
 }
